requirements.txt
cryptography==41.0.1
defusedxml==0.7.1
cffi==1.15.1
future==0.18.3
Pillow==9.5.0
pip==23.1.2
attrs==23.1.0
argparse==1.4.0
wheel==0.40.0
Django==4.2.2
django-axes==6.0.4
requests==2.31.0
build==0.10.0
virtualenv==20.23.1
yarg==0.1.9
setuptools==68.0.0
asgiref==3.7.2
sqlparse==0.4.4
psycopg2==2.9.6
colorama==0.4.6
click==8.1.3
django-allauth==0.54.0
PyJWT==2.7.0
certifi==2023.5.7
docopt==0.6.2
pipreqs==0.4.13
six==1.16.0
oauthlib==3.2.2
idna==3.4
urllib3==2.0.3
pycparser==2.21
packaging==23.1
distlib==0.3.6
filelock==3.12.2
platformdirs==3.8.0
django-embed-video==1.4.8
django-model-utils==4.3.1
python-dotenv==1.0.0